Refactor: Introduce JellyfinApiClient
This class will handle creating Api instances with the correct config,
base url and axios instance. This is important so that the apis use
the axios instance we depend on, rather than @jellyfin/axios-client
as if they use the copy from @jellyfin/axios-client, our mocks will
not apply in testing.